Add Jenkinsfile for bundle build
I've added in a separate Jenkinsfile to create the lambda bundle, upload
it to S3 and deploy the lambda.
This needs to be called twice, once in the test multi branch job for the
master branch and once in the build job which updates the virus
definitions so it makes sense to have it in a separate job.