Remove Yara-Rules rule source
I added this in without thinking about it and it produces matches for
all files including ones that definitely aren't viruses and also, it
stops the antivirus code working at all.

At some point, we need to have a proper investigation as to which yara
files we use. The signature-base one is the only one airbnb use so we
might be ok just with that but it's worth checking.

In the meantime, to get the antivirus working, this change will do it.