Fix order of docker build and push steps in Jenkinsfile-build
Push each image as it is built, rather than after building them all.
Some of these images depend on ones that are built earlier in the
script. Building them all together used to work, but it doesn't any more
because we added the `--pull` flag, which means the base image has to
be pushed to the repo so that we can pull it.