Use requirements file to install runtime Python dependencies
Before, the requirements file was used to run the tests, but it
wasn't used in the Docker container, so the runtime dependencies
had to be specified separately in the script run by the Dockerfile.

This change commonises the two, so the runtime dependencies are only
specified in one place (requirements-runtime.txt).

This makes it easier to maintain the list of dependencies, but it
also has a security benefit, because Snyk notifies us if there are
any vulnerabilities in the dependencies listed in the requirements
files.