Add script to exclude rules based on the rules json.
This uses the plyara library to parse the rules files specified in the
rules json. The parsed rules are then filtered to exlude any rule names
specified in rules.json and the filtered rules are written back to the
file. This allows us to exclude a single rule if it's causing us
problems.