Simplify the antivirus build process
The build process before was overly complicated so I've simplified this
down to a Jenkinsfile-test and Jenkinsfile-deploy like our other
projects.

The Jenkinsfile-test for master builds will build all of the yara docker
images from scratch and use them to build the lambda. It will not push
them and I'll delete the repositories from ECR. If we're not storing the
images we use to build the lambda, it removes the confusion around the
version tags for those images.

This takes around 10 minutes for a full build of the antivirus but we
don't change it that often so this seems fine.