Build the docker image on every deploy.
Before, we built the images which compiled the rules and installed
dependencies then pushed them to ECR and then pulled them down to build
the lambda but this is causing confusion with the versioning of the
lambda code and the version of these images.

This change builds the docker images on each deploy, runs them to build
the lambda and doesn't push them.