Reset message visibility and delete successful messages on failure
The original antivirus lambda didn't have any code to delete successful
mesages when one message in a batch failed. This hasn't been a problem
because the lambda has been getting one message at a time but it's worth
adding in here to bring it into line with the other lambdas.

I've also added in a call to set the message visibility to zero on error
so the message can be picked up again immediately.