Remove cognito id and get dirty bucket name from input.
We are changing the download files lambda to send userId instead of
cognitoId. The values will be the same but it's no longer a cognito ID
so the name is changing to avoid confusion.

We also need to pass in the dirty bucket name because we'll need the
antivirus check to work with both the existing dirty bucket and the new
cloudwatch bucket.

I've also updated the README instructions.