Log the time taken as json.
This allows the performance checks to work out how long each lambda has
taken to run.

We may move this to the database in the future but this works for now.

This is python so instead of configuring the standard logger which I
don't think works, we're just printing the json string to stdout. This
does work.