Use updated openssl
The gcc64 gcc64-c++ packages don't exist and don't do anything. I've removed them

The version of the openssl on the amazonlinux image is old and doesn't have a lot of the functions that yara expects.
You can install a newer version in the openssl11 package which has the functions but is called openssl11 so yara can't find it.
I've changed it so the configure step looks in the new openssl11 package and this seems to work.

The lambda was complaining about libcrypto.so.1.1 This is in the openssl11-static package which is now installed and copied over to the lambda directory.